fetchDenoDeps: improve unit tests
