+++ pkgs/build-support/deno/fetch-deno-deps/scripts/deno/src/lockfile-transformer/lockfile-transformer.test.ts
@@ -1,24 +1,6 @@
 import { parsePackageSpecifier } from "./lockfile-transformer.ts";
 import type { PackageSpecifier } from "../types.d.ts";
-
-export function assertEq(a: any, b: any, msg?: string) {
-  function throwNow() {
-    const aString = typeof a === "object" ? JSON.stringify(a, null, 2) : a;
-    const bString = typeof b === "object" ? JSON.stringify(b, null, 2) : b;
-    const _msg = `Assertion failed:
-"${aString}"
-  ===
-"${bString}"
-`;
-    throw new Error(_msg + (msg || ""));
-  }
-  if (JSON.stringify(a) === JSON.stringify(b)) {
-    return;
-  } else if (a === b) {
-    return;
-  }
-  throwNow();
-}
+import { assertEq } from "../utils.ts";
 
 Deno.test("test parsePackageSpecifier", () => {
   type TestCase = {
